requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.5.4=pypi_0
async-timeout=3.0.1=pypi_0
attrs=19.3.0=pypi_0
ca-certificates=2020.12.8=haa95532_0
cachecontrol=0.12.5=pypi_0
cachetools=3.1.1=pypi_0
certifi=2019.9.11=pypi_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
discord-py=1.2.4=pypi_0
firebase-admin=3.1.0=pypi_0
flask=1.1.1=pypi_0
google-api-core=1.14.3=pypi_0
google-api-python-client=1.7.11=pypi_0
google-auth=1.6.3=pypi_0
google-auth-httplib2=0.0.3=pypi_0
google-cloud-core=1.0.3=pypi_0
google-cloud-firestore=1.5.0=pypi_0
google-cloud-storage=1.21.0=pypi_0
google-resumable-media=0.4.1=pypi_0
googleapis-common-protos=1.6.0=pypi_0
grpcio=1.24.3=pypi_0
httplib2=0.14.0=pypi_0
idna=2.8=pypi_0
idna-ssl=1.1.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.10.3=pypi_0
kwconfig=1.1.7=pypi_0
markupsafe=1.1.1=pypi_0
msgpack=0.6.2=pypi_0
multidict=4.5.2=pypi_0
openssl=1.1.1i=h2bbff1b_0
pip=20.3.3=py37haa95532_0
protobuf=3.10.0=pypi_0
psycopg2-binary=2.8.4=pypi_0
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
python=3.7.3=h8c8aaf0_1
python-dotenv=0.15.0=pypi_0
python-firebase=1.2=pypi_0
pytz=2019.3=pypi_0
requests=2.22.0=pypi_0
rsa=4.0=pypi_0
search-google=1.2.1=pypi_0
setuptools=51.0.0=py37haa95532_2
six=1.12.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
typing-extensions=3.7.4.1=pypi_0
tzdata=2020d=h14c3975_0
uritemplate=3.0.0=pypi_0
urllib3=1.25.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
websockets=6.0=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py37_0
yarl=1.3.0=pypi_0
zlib=1.2.11=h62dcd97_4